requirements.txt
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
marshmallow==3.0.0rc5
onetimepass==1.0.1
PyJWT==1.7.1
PyMySQL==0.9.3
PyQRCode==1.2.1
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.24.2
Werkzeug==0.15.6
WTForms==2.2.1
flask_redis==0.3.0
